feat(api): split auxiliary routes into modules
